delete run services asynchronously so it finishes faster

+++ scripts/ci_scripts/cleanup_dev_functions.sh
@@ -10,16 +10,24 @@ current_run_services=$(gcloud run services list --format="value(name)")
 to_delete_functions=$(grep -Fxvf <(echo "$active_branches") <(echo "$current_functions") || echo "")
 to_delete_services=$(grep -Fxvf <(echo "$active_branches" | sed 's/_/-/g') <(echo "$current_run_services") || echo "")
 
-echo "start deleting functions"
-while IFS= read -r branch_name; do
-    echo $branch_name
-    gcloud functions delete $branch_name --quiet --region=us-central1 --project=ib-serverless-api-dev
-done <<< "$to_delete_functions"
-echo "finish deleting functions"
+if [[ ! -z "$to_delete_functions" ]]; then
+    echo "start deleting functions"
+    while IFS= read -r branch_name; do
+        echo $branch_name
+        gcloud functions delete $branch_name --quiet --region=us-central1 --project=ib-serverless-api-dev
+    done <<< "$to_delete_functions"
+    echo "finish deleting functions"
+else
+    echo "no functions to delete"
+fi
 
-echo "start deleting run services"
-while IFS= read -r branch_name; do
-    echo $branch_name
-    gcloud run services delete $branch_name --quiet --region=us-central1 --project=ib-serverless-api-dev --namespace=961991692823
-done <<< "$to_delete_services"
-echo "finish deleting run services"
+if [[ ! -z "$to_delete_functions" ]]; then
+    echo "start deleting run services"
+    while IFS= read -r branch_name; do
+        echo $branch_name
+        gcloud run services delete $branch_name --quiet --region=us-central1 --project=ib-serverless-api-dev --namespace=961991692823 --async
+    done <<< "$to_delete_services"
+    echo "finish deleting run services"
+else
+    echo "no run services to delete"
+fi
